% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcrules.glpk-package.R
\docType{package}
\name{gcrules.glpk-package}
\alias{gcrules.glpk}
\alias{gcrules.glpk-package}
\title{gcrules.glpk: GLPK backend for Globally Consistent Rules (gcrules) package}
\description{
GLPK backend for Globally Consistent Rules (gcrules) package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamorarr/gcrules.glpk}
  \item Report bugs at \url{https://github.com/zamorarr/gcrules.glpk/issues}
}

}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
