% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-support.r
\name{max_support}
\alias{max_support}
\title{Find Max Support for a Binary Matrix}
\usage{
max_support(x, y, xe, ye, R_max = ncol(x), ws = 1, complement = TRUE)
}
\arguments{
\item{x}{binary matrix of features}

\item{y}{binary vector of predictor}

\item{xe}{binary vector of example observation}

\item{ye}{binary value of example predictor}

\item{complement}{whether to add the complement of x as features}
}
\description{
Find Max Support for a Binary Matrix
}
