% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-set-cover.r
\name{min_set_cover}
\alias{min_set_cover}
\title{Minimum Set Cover for a Binary Matrix}
\usage{
min_set_cover(x, y, xe, ye, complement = TRUE)
}
\arguments{
\item{x}{binary matrix of features}

\item{y}{binary vector of predictor}

\item{xe}{binary vector of example observation}

\item{ye}{binary value of example predictor}

\item{complement}{whether to add the complement of x as features}
}
\description{
Minimum Set Cover for a Binary Matrix
}
