% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcrules.glpk-package.R
\docType{package}
\name{gcrules.glpk-package}
\alias{gcrules.glpk}
\alias{gcrules.glpk-package}
\title{gcrules.glpk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
